+++ examples/examples_yaml_test.go
@@ -474,6 +474,18 @@ func TestYamlInsightsAccountExample(t *testing.T) {
 	runPulumiTest(t, test)
 }
 
+func TestYamlAdoIntegrationExample(t *testing.T) {
+	cwd := getCwd(t)
+	integration.ProgramTest(t, &integration.ProgramTestOptions{
+		Dir: path.Join(cwd, ".", "yaml-ado-integration"),
+		Config: map[string]string{
+			"organizationName":    getOrgName(),
+			"adoOrganizationName": "test-ado-org",
+			"projectId":           "test-project-id",
+		},
+	})
+}
+
 func writePulumiYaml(t *testing.T, yamlContents interface{}) string {
 	tmpdir := t.TempDir()
 	b, err := yaml.Marshal(yamlContents)